[CCUBE-2183][GZ] consolidate services
